refactor(app): #373 Phase 1-7 use-claude-check.ts に CLI 検査を切り出し
App.tsx 1468 -> 1430 行 (-38)。Claude CLI 検査 (claudeCheck state +
runClaudeCheck callback + claudeCommand 変更検知 effect) と起動時
アップデーター遅延 effect (5 秒後の silentCheckForUpdate) を新規 hook
src/renderer/src/lib/hooks/use-claude-check.ts に集約。

同居の根拠: tasks/refactor-handoff.md で Phase 1-7 として
「claudeCheck / アップデーター遅延 effect」が一括指定されており、
両者とも「起動時 1 回 / 設定変更時に追従する App ライフサイクル系の
副作用」という共通点を持つ。内部実装は別 useEffect で完全分離。

Phase 1-1 〜 1-6 と揃えた点:
- opts なし純粋 hook (Phase 1-5 use-layout-resize と同じ流儀)
- useSettingsValue('claudeCommand') を hook 内で直接呼ぶ
- 戻り値は { claudeCheck, runClaudeCheck } の最小限
- 派生 flag (claudeReady / claudeMissing 等) は作らない: JSX 側で
  claudeCheck.state === 'X' を直接判定 (現状コードと完全一致)

不変式 (現状挙動を完全保持):
- runClaudeCheck の deps は [claudeCommand] のまま (identity 安定化は
  別 issue 候補)
- 再検査 effect の deps は [runClaudeCheck] のまま
- アップデーター effect の cancelled flag + clearTimeout race 防止を保持
- window.api.app.checkClaude(claudeCommand || 'claude') の '|| claude'
  フォールバック保持
- dynamic import パスは ./lib/updater-check → ../updater-check に
  hook ファイル位置に合わせて訂正 (機能は同一)

検証:
- npm run typecheck: green
- cargo check --manifest-path src-tauri/Cargo.toml: green
- cargo clippy --no-deps: 既存ベースライン 15 件のまま (新規警告ゼロ)

累計削減 (Phase 1-1 〜 1-7): -1397 行 (#373 全体目標 -1995, 進捗 70%)。

Refs #373
